refactor(evals): co-locate pointcloud-only generators and scorers under suites/pointcloud/lib
The shared row reader keeps the numeric and mcq types and the plain window
context, and takes grade_of/select_of hooks. The point-cloud suites' lib
adds the coords type, the fused context entry, the cell-grid helpers, the
eight row generators, and the coord_list/matched_set scorers, with their
tests beside them.
