+++ dimos/evals/suites/pointcloud/go2_pointcloud_free_disk.py
@@ -44,7 +44,7 @@ from pathlib import Path
 import numpy as np
 from scipy import ndimage
 
-from dimos.evals.suites.lib import generate
+from dimos.evals.suites.pointcloud.lib import generate
 from dimos.evals.types import Suite
 
 _JSON = Path(__file__).parent / "go2_pointcloud_free_disk.json"
