+++ dimos/evals/suites/pointcloud/go2_pointcloud_rooms.py
@@ -45,7 +45,7 @@ from __future__ import annotations
 import json
 from pathlib import Path
 
-from dimos.evals.suites.lib import generate
+from dimos.evals.suites.pointcloud.lib import generate
 from dimos.evals.types import Suite
 
 _JSON = Path(__file__).parent / "go2_pointcloud_rooms.json"
